collections -> collection.abc
collections.abc was added under collections, and the use of collections
for abstract base classes without collectsion.abc was deprecated in
Python 3.3. It was supposedly going to be removed in Python 3.8, but I'm
not sure it actually got removed that early. In any case, reports are
coming in for missing collections.Iterable with people running Python
3.10. This updates to the use of collections.abc.